refactor: switch to nuxtui horizontal navigation component for tabs
